otp: Add DEVELOPMENT howto and complete `make test` support
This commit polishes a lot of the make system to be more
robust and hopefully user friendly. A new DEVELOPMENT howto
guide has been added with instructions on how to work with
the Erlang/OTP code and the `make test` functionality has
been fixed so that it should work for all applications.

In order to solve the problem of some applications relying
on being tested using a release, each application has been
instrumented in its Makefile so that it knows if a release
has to be done or not.