ct: Do not flush output when exiting ct_run
When running tests the state of the system may not always
be able to flush all ports and we do not want the system
to hang when that happens.

An example of this happening is when the peer_SUITE has
connections to dead nodes that have been been disconnected
yet.